RAG_Based_AI_Assistant/requirements.txt
openai-whisper==20231117
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
